# Conflicts:
#	src/main/java/com/example/nanuer_server/domain/entity/MyPageEntity.java
#	src/main/java/com/example/nanuer_server/domain/entity/PostEntity.java
#	src/main/java/com/example/nanuer_server/domain/entity/UserEntity.java
#	src/main/java/com/example/nanuer_server/domain/repository/PostRepository.java
#	src/main/java/com/example/nanuer_server/domain/repository/UserRepository.java

+++ src/main/java/com/example/nanuer_server/domain/repository/PostRepository.java
@@ -3,8 +3,5 @@ package com.example.nanuer_server.domain.repository;
 import com.example.nanuer_server.domain.entity.PostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostRepository extends JpaRepository<PostEntity, Integer> {
-
-
-
+public interface PostRepository extends JpaRepository<PostEntity,Long> {
 }
